requirements/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
beautifulsoup4==4.12.3
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
debugpy==1.8.5
decorator==5.1.1
easyocr==1.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
fsspec==2024.6.1
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
imageio==2.35.1
imgaug==0.4.0
imutils==0.5.4
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==3.20.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.26.1
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pypi==2.1
pytesseract==0.3.13
pytest==8.3.2
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
setuptools==73.0.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.2
termcolor==2.4.0
textblob==0.18.0.post0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.8.10
torchvision==0.19.0
tqdm==4.66.5
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wheel==0.44.0
wordcloud==1.9.3
wrapt==1.16.0
